requirements.txt
#Dependências
mysql-connector-python==8.3.0
pytest==8.2.0